refactor: split doctor.rs god-file into modules
