+++ src/commands/runner/doctor/tests/mod.rs
@@ -1,0 +1,8 @@
+mod browser;
+mod daemon;
+mod extension_parity;
+mod homeboy_path;
+mod managed_source;
+mod provider;
+mod shape;
+mod tools;
